Some update.

+++ src_number/NumberTheoryBoostCppInt.h
@@ -857,11 +857,11 @@ inline void NearestInteger(boost::multiprecision::cpp_rational const &xI,
 }
 
 inline void set_to_infinity(boost::multiprecision::cpp_rational &x) {
-  x = std::numeric_limits<size_t>::max();
+  x = std::numeric_limits<uint64_t>::max();
 }
 
 inline void set_to_infinity(boost::multiprecision::cpp_int &x) {
-  x = std::numeric_limits<size_t>::max();
+  x = std::numeric_limits<uint64_t>::max();
 }
 
 inline boost::multiprecision::cpp_rational
